requirements.txt
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
rsa==4.9
setuptools==75.1.0
Telethon==1.36.0
Unidecode==1.3.8
wheel==0.44.0
